LU-10391 ptlrpc: change cc_nid in nrs to be struct lnet_nid
switch to struct lnet_nid and adjust accordingly.

Note the return value of nrs_crrn_cmpfn() is changed from a bool to
"0 or negative error" to make it more clear which is success and which
is failure.  The end result is the same.

Test-Parameters:trivial testlist=sanityn envdefinitions=ONLY=77b
Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I3e5f2e9d88d7aeccb68ebfb4c814cc4e388cbc42
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/50100
Reviewed-by: Chris Horn <chris.horn@hpe.com>
Reviewed-by: jsimmons <jsimmons@infradead.org>
Reviewed-by: Nikitas Angelinas <nikitas.angelinas@hpe.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>