LU-15748 ofd: fix fallocate interop for older clients
The logic for detecting older client fallocate was backward, and
should be checking if the OBD_CONNECT_OLD_FALLOC patch was *not*
present to detect old clients.

Since the new server does not have OLD_FALLOC in the SUPPORTED flags,
it will clear the flag from the export at connect time, so it needs
to be saved in the export early on.

Test-Parameters: testlist=sanityn env=ONLY=16,HONOR_EXCEPT=y
Test-Parameters: serverversion=2.14 testlist=sanityn env=ONLY=16,HONOR_EXCEPT=y
Fixes: 7905359296 ("LU-15748 osc: fallocate interop for 2.14 clients")
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I327183025a8de6fd814a7c2929365497153ebbe5
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/47548
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>