LU-12511 test: use proper UAPI header for fsx test app
Using uapi as part of the path of the header forces the use of
the OpenSFS tree internal headers instead of an possible
external UAPI headers; which is the case for the native Linux
Lustre client. Just use linux/lustre path for UAPI headers
which works in all cases.

Change-Id: Ia46021f39a8f2e4ecae71846cbb76464617d12ab
Test-Parameters: trivial
Fixes: 90ba8b4ac ("LU-14642 test: add fsx mirror file test mode")
Signed-off-by: James Simmons <jsimmons@infradead.org>
Reviewed-on: https://review.whamcloud.com/47565
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Bobi Jam <bobijam@hotmail.com>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>