LU-16262 tests: Remove sanity-gss.sh
Purpose of sanity-gss is to test just the
GSSAPI code itself. This is done by making
use of the gssnull/null security flavor.
Currently, this is exercised through
non-regression tests with SSK and therefore
sanity-gss.sh is not required. This patch
removes sanity-gss.sh from repo due to above
reasons.

Signed-off-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Change-Id: Ief62e8111cafdc5bebca1f47a1b09fbafb152a76
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/48948
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>