LU-8837 ptlrpc: fix problem with distclean
"make distclean" follows the filesystem hierarchy so ldlm/Makefile
will be cleaned before ptlrpc/Makefile is read, which triggers an
error now that ptlrpc/Makefile tries to include ldlm/Makefile

This is an uninteresting error as we don't need to include
ldlm/Makefile in the distclean case.

So replace the "include" with "-include" which suppresses the error.

Fixes: f32bdde0574d ("LU-8837 lustre: make ldlm and target file lists")
Test-Parameters: trivial
Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I66c4552b3806113450f9636b7916aff7ed69bda3
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/49126
Tested-by: James Simmons <jsimmons@infradead.org>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>