LU-16382 spec: SUSE OBS requires kernel-source requirement.
The SUSE OBS creates a virtual environment containing
ONLY the stated BuildRequires requirements, and some defaults.
To be able to build ldiskfs we need the kernel source, so we
need "BuildRequires: kernel-source".

However when contrib/lbuild it is used to build lustre, it finds the
source by other means and fails on that BuildRequires.  So it must be
conditional on running under abuild. (the OBS build tool).

When abuild extracts these BuildRequires, it cannot parse %() so
conditions using that all appear to be "false".  But %() is the only
way to detect abuild - but looking in environment.

So a dance is needed to fit with all these odd requirements.

Test-Parameters: trivial
Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I2fe9ecaf857ecbd5fda7e857b661b5b756501190
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/49365
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>